fix: 결과 출력 수정
- 경주 전 초기값 출력하도록 수정
- 입력 받은 값 List 객체로 리턴하도록 변경
