+++ src/main/java/com/kimdahyeee/racing/view/InputView.java
@@ -1,14 +1,26 @@
 package com.kimdahyeee.racing.view;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
-    public static String getCarNames() {
+    private static final String NAME_SEPARATOR = ",";
+
+    public static List<String> getCarNames() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         String names = scanner.nextLine();
 
-        return names;
+        return getCarNames(names);
+    }
+
+    private static List<String> getCarNames(String names) {
+        if (names == null || names.trim().isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        return Arrays.asList(names.split(NAME_SEPARATOR));
     }
 
     public static int getTryCount() {
